Reorganize files and change imports
