+++ packages/extension/src/view/devtools/pages/privateAdvertising/protectedAudience/auctions/panel.tsx
