--- packages/extension/src/view/devtools/pages/privateAdvertising/protectedAudience/auctions/v2/index.tsx
@@ -1,41 +1,0 @@
-/*
- * Copyright 2024 Google LLC
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     https://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-/**
- * External dependencies.
- */
-import React from 'react';
-import { SidebarProvider } from '@google-psat/design-system';
-
-/**
- * Internal dependencies.
- */
-import useSidebarProcessing from './hooks/useSidebarProcessing';
-import Panel from './panel';
-
-const AuctionsV2 = () => {
-  const { sidebarData } = useSidebarProcessing();
-
-  return (
-    <SidebarProvider
-      data={sidebarData}
-      defaultSelectedItemKey={Object.keys(sidebarData)[0]}
-    >
-      <Panel />
-    </SidebarProvider>
-  );
-};
-
-export default AuctionsV2;
