+++ packages/extension/src/view/devtools/pages/privateAdvertising/protectedAudience/explorableExplanation/tableTabPanels/auctions/index.tsx
@@ -28,9 +28,9 @@ import type {
 /**
  * Internal dependencies.
  */
-import AdUnitsPanel from '../../adUnits/panel';
-import type { AuctionEventsType } from '../../../../../stateProviders/protectedAudience/context';
-import AuctionsContainer from '../../auctions/v1/container';
+import AdUnitsPanel from '../../../adUnits/panel';
+import type { AuctionEventsType } from '../../../../../../stateProviders/protectedAudience/context';
+import AuctionsContainer from './container';
 
 interface AuctionsProps {
   auctionEvents: {
